+++ tkeron.d.ts
@@ -1,1 +1,2 @@
 declare const com: HTMLElement;
+declare const document: Document;